+++ sembast/lib/src/filter_impl.dart
@@ -43,6 +43,9 @@ class SembastCustomFilter extends SembastFilterBase {
       return false;
     }
   }
+
+  @override
+  String toString() => 'SembastCustomFilter()';
 }
 
 abstract class _FilterAnyInList {