+++ sembast/lib/src/query_ref_impl.dart
@@ -1,3 +1,4 @@
+import 'package:meta/meta.dart';
 import 'package:sembast/src/common_import.dart';
 import 'package:sembast/src/database_impl.dart';
 import 'package:sembast/src/debug_utils.dart';
@@ -13,7 +14,6 @@ class SembastQueryRef<K, V> implements QueryRef<K, V> {
   final StoreRef<K, V> store;
 
   /// The finder.
-  // ignore: deprecated_member_use_from_same_package
   final SembastFinder? finder;
 
   /// Query ref implementation.
@@ -26,8 +26,26 @@ class SembastQueryRef<K, V> implements QueryRef<K, V> {
 
   @override
   String toString() => '$store $finder)';
+}
 
-  @override
+/// Internal access
+@protected
+extension SembastQueryRefExtensionPrv<K, V> on QueryRef<K, V> {
+  /// Sembast query ref. implementation.
+  SembastQueryRef<K, V> get sembastQueryRef => this as SembastQueryRef<K, V>;
+}
+
+/// Common extension
+extension SembastQueryRefCommonExtension<K, V> on QueryRef<K, V> {
+  /// Finder if any.
+  Finder? get finder => sembastQueryRef.finder;
+}
+
+/// Query db actions.
+extension SembastQueryRefExtension<K, V> on QueryRef<K, V> {
+  /// Find multiple records and listen for changes.
+  ///
+  /// Returns a single subscriber stream that must be cancelled.
   Stream<List<RecordSnapshot<K, V>>> onSnapshots(Database database) {
     var db = getDatabase(database);
     // Create the query but don't add it until first result is set
@@ -41,9 +59,10 @@ class SembastQueryRef<K, V> implements QueryRef<K, V> {
       try {
         await db.notificationLock.synchronized(() async {
           // Find all matching, ignoring offset/limit but order them
-          var allMatching = await (store as SembastStoreRef<K, V>)
-              .findImmutableRecords(database,
-                  finder: finder?.cloneWithoutLimits());
+          var allMatching = await sembastQueryRef.store.findImmutableRecords(
+              database,
+              finder: sembastQueryRef.finder?.cloneWithoutLimits()
+                  as SembastFinder?);
           // ignore: unawaited_futures
 
           // Get the result at query time first
@@ -60,21 +79,55 @@ class SembastQueryRef<K, V> implements QueryRef<K, V> {
     return ctlr.stream;
   }
 
-  @override
+  /// Find multiple records.
+  ///
+  /// Returns an empty array if none found.
   Future<List<RecordSnapshot<K, V>>> getSnapshots(DatabaseClient client) =>
-      store.find(client, finder: finder);
+      sembastQueryRef.store.find(client, finder: sembastQueryRef.finder);
 
-  @override
+  /// Find first record matching the query.
+  ///
+  /// Returns null if none found.
   Future<RecordSnapshot<K, V>?> getSnapshot(DatabaseClient client) =>
-      store.findFirst(client, finder: finder);
+      sembastQueryRef.store.findFirst(client, finder: sembastQueryRef.finder);
 
-  @override
+  /// Find first record (null if none) and listen for changes.
+  ///
+  /// Returns a single subscriber stream that must be cancelled.
   Stream<RecordSnapshot<K, V>?> onSnapshot(Database database) {
-    if (finder?.limit != 1) {
-      return SembastQueryRef(store, cloneFinderFindFirst(finder))
+    if (sembastQueryRef.finder?.limit != 1) {
+      return SembastQueryRef(sembastQueryRef.store,
+              cloneFinderFindFirst(sembastQueryRef.finder))
           .onSnapshot(database);
     }
     return onSnapshots(database)
         .map((list) => list.isNotEmpty ? list.first : null);
   }
+
+  /// count records.
+  Future<int> count(DatabaseClient client) async =>
+      (await getSnapshots(client)).length;
+
+  /// onCount stream, called when the number of items changes.
+  Stream<int> onCount(Database database) =>
+      onSnapshots(database).map((e) => e.length);
+}
+
+/// Query db actions. synchronous access.
+extension SembastQueryRefSyncExtension<K, V> on QueryRef<K, V> {
+  /// Find multiple records. Synchronous version.
+  ///
+  /// Returns an empty array if none found.
+  List<RecordSnapshot<K, V>> getSnapshotsSync(DatabaseClient client) =>
+      sembastQueryRef.store.findSync(client, finder: sembastQueryRef.finder);
+
+  /// Find first record matching the query. Synchrnous version.
+  ///
+  /// Returns null if none found.
+  RecordSnapshot<K, V>? getSnapshotSync(DatabaseClient client) =>
+      sembastQueryRef.store
+          .findFirstSync(client, finder: sembastQueryRef.finder);
+
+  /// count records. Synchronous version.
+  int countSync(DatabaseClient client) => getSnapshotsSync(client).length;
 }